updated a few endpoints after testing them. Needs more testing, but they 'work'

+++ backend/src/routes/notification.ts
@@ -1,15 +1,16 @@
 import { Request, Response, Router } from "express";
 import { authMW } from "../middleware/auth.ts";
-import { NotificationModel, UserModel } from "src/models/exports.ts";
+import { NotificationModel, UserModel } from "../models/exports.ts";
 
 const notificationRead = async (req: Request, res: Response) => {
 	try {
 		const usersub = res.locals.usersub;
 		const noteId = req.params.noteId;
 		const user = await UserModel.findOne({ googleId: usersub });
-		const note = user?.notifications.find((notification) => {
+		const note = user?.notifications.filter((notification) => {
 			return notification.id == noteId;
 		})
+		console.log(note)
 		if (!note) {
 			res.statusCode = 500;
 			res.send();
@@ -26,16 +27,19 @@ const notificationDelete = async (req: Request, res: Response) => {
 	try {
 		const usersub = res.locals.usersub;
 		const noteId = req.params.noteId;
-		const user = await UserModel.findOneAndUpdate({ googleId: usersub }, { $pull: { notifications: { id: noteId } } });
-		const note = user?.notifications.find((notification) => {
-			return notification.id == noteId;
-		})
-		if (note) {
-			console.error(note);
-			res.statusCode = 500;
-			res.send();
+		const user = await UserModel.findOneAndUpdate({ googleId: usersub, notifications: noteId }, { $pull: { notifications: { id: noteId } } });
+		if (user) {
+			const note = user?.notifications.filter((notification) => {
+				return notification.id == noteId;
+			})
+			if (note) {
+				console.error(note);
+				res.statusCode = 500;
+				res.send();
+			}
+			// Only delete notification if a user was found
+			await NotificationModel.findByIdAndDelete(noteId);
 		}
-		await NotificationModel.findByIdAndDelete(noteId);
 		res.send();
 	} catch (err) {
 		res.statusCode = 500;
@@ -43,8 +47,8 @@ const notificationDelete = async (req: Request, res: Response) => {
 	}
 };
 
-export const notificationRouter = Router();
-notificationRouter.use(authMW);
+export const noteRouter = Router();
+noteRouter.use(authMW);
 
-notificationRouter.put("/read/:noteId", notificationRead);
-notificationRouter.delete("/delete/:noteId", notificationDelete);
+noteRouter.put("/read/:noteId", notificationRead);
+noteRouter.delete("/delete/:noteId", notificationDelete);
